+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/assets/TableAssetSourcesResource.java
@@ -20,6 +20,12 @@ public class TableAssetSourcesResource {
     return tableAssetSourcesService.getAssetSources();
   }
 
+
+  @GetMapping("/default")
+  public TableAssetSource getDefaultSource() {
+    return tableAssetSourcesService.getDefaultAssetSource();
+  }
+
   @GetMapping("/{sourceId}")
   public TableAssetSource getSource(@PathVariable("sourceId") String sourceId) {
     return tableAssetSourcesService.getAssetSource(sourceId);